opaque: isolate shadowed error test int oseparate testdata package
