+++ opaque/opaque_test.go
@@ -18,5 +18,6 @@ func TestAnalyzer(t *testing.T) {
 	analysistest.RunWithSuggestedFixes(t, testdata, opaque.Analyzer, "g")
 	analysistest.RunWithSuggestedFixes(t, testdata, opaque.Analyzer, "h")
 	analysistest.RunWithSuggestedFixes(t, testdata, opaque.Analyzer, "i")
+	analysistest.RunWithSuggestedFixes(t, testdata, opaque.Analyzer, "j")
 	analysistest.RunWithSuggestedFixes(t, testdata, opaque.Analyzer, "x")
 }
