+++ opaque/testdata/src/i/i.go.golden
@@ -24,20 +24,6 @@ func WithError() (*myInterfaceImpl, error) { // want "'WithError' function retur
 	return &myInterfaceImpl{}, nil
 }
 
-type myError struct{}
-
-func (e *myError) Error() string {
-	return "error"
-}
-
-type error interface { // this shadows the predeclared error
-	Error() string
-}
-
-func GetError() *myError { // want "'GetError' function return 'error' interface at the 1st result, abstract a single concrete implementation of '\\*myError'"
-	return &myError{}
-}
-
 type MyError interface {
 	Error() string
 }
